Добавлен раздел управления горящими предложениями

+++ scripts/admin/news/ajaxEditNews.php
@@ -18,50 +18,54 @@ $url = $mysqli->real_escape_string($_POST['url']);
 $description = $mysqli->real_escape_string(nl2br($_POST['description']));
 $text = $mysqli->real_escape_string(nl2br($_POST['text']));
 
-$urlCheckResult = $mysqli->query("SELECT COUNT(id) FROM ft_news WHERE url = '".$url."' AND id <> '".$id."'");
-$urlCheck = $urlCheckResult->fetch_array(MYSQLI_NUM);
+if(!is_numeric($url)) {
+    $urlCheckResult = $mysqli->query("SELECT COUNT(id) FROM ft_news WHERE url = '".$url."' AND id <> '".$id."'");
+    $urlCheck = $urlCheckResult->fetch_array(MYSQLI_NUM);
 
-if($urlCheck[0] == 0) {
-    $newsResult = $mysqli->query("SELECT * FROM ft_news WHERE id = '".$id."'");
-    $news = $newsResult->fetch_assoc();
+    if($urlCheck[0] == 0) {
+        $newsResult = $mysqli->query("SELECT * FROM ft_news WHERE id = '".$id."'");
+        $news = $newsResult->fetch_assoc();
 
-    if(!empty($_FILES['preview']['name'])) {
-        if($_FILES['preview']['error'] == 0 and substr($_FILES['preview']['type'], 0, 5) == "image") {
-            $previewTmpName = $_FILES['preview']['tmp_name'];
-            $previewName = randomName($previewTmpName);
-			$previewDBName = $previewName.".".substr($_FILES['preview']['name'], count($_FILES['preview']['name']) - 4, 4);
-			$previewUploadDir = "../../../img/news/preview/";
-			$previewUpload = $previewUploadDir.$previewDBName;
+        if(!empty($_FILES['preview']['name'])) {
+            if($_FILES['preview']['error'] == 0 and substr($_FILES['preview']['type'], 0, 5) == "image") {
+                $previewTmpName = $_FILES['preview']['tmp_name'];
+                $previewName = randomName($previewTmpName);
+                $previewDBName = $previewName.".".substr($_FILES['preview']['name'], count($_FILES['preview']['name']) - 4, 4);
+                $previewUploadDir = "../../../img/news/preview/";
+                $previewUpload = $previewUploadDir.$previewDBName;
 
-			if($mysqli->query("UPDATE ft_news SET preview = '".$previewDBName."' WHERE id = '".$id."'")) {
-			     unlink($previewUploadDir.$news['preview']);
-			     resize($previewTmpName, 270);
-			     move_uploaded_file($previewTmpName, $previewUpload);
+                if($mysqli->query("UPDATE ft_news SET preview = '".$previewDBName."' WHERE id = '".$id."'")) {
+                     unlink($previewUploadDir.$news['preview']);
+                     resize($previewTmpName, 270);
+                     move_uploaded_file($previewTmpName, $previewUpload);
+                } else {
+                    echo "upload";
+
+                    $req = ob_get_contents();
+                    ob_end_clean();
+                    echo json_encode($req);
+
+                    exit;
+                }
             } else {
-			    echo "upload";
-
-			    $req = ob_get_contents();
-                ob_end_clean();
-                echo json_encode($req);
-
-                exit;
+                echo "preview";
             }
-        } else {
-            echo "preview";
         }
-    }
 
-    $description = str_replace("<br>", " ", $description);
-    $description = str_replace("<br/>", " ", $description);
-    $description = str_replace("<br />", " ", $description);
+        $description = str_replace("<br>", " ", $description);
+        $description = str_replace("<br/>", " ", $description);
+        $description = str_replace("<br />", " ", $description);
 
-    if($mysqli->query("UPDATE ft_news SET header = '".$header."', description = '".$description."', text = '".$text."', url = '".$url."' WHERE id = '".$id."'")) {
-        echo "ok";
+        if($mysqli->query("UPDATE ft_news SET header = '".$header."', description = '".$description."', text = '".$text."', url = '".$url."' WHERE id = '".$id."'")) {
+            echo "ok";
+        } else {
+            echo "failed";
+        }
     } else {
-        echo "failed";
+        echo "url";
     }
 } else {
-    echo "url";
+    echo "number";
 }
 
 $req = ob_get_contents();
